HAL.AspNetCore 7.1.0 HAL.AspNetCore.OData 6.1.0 HAL.Common 2.1.0
Better support for Nullable Reference Types in HAL-Forms.
Default Form-Templates now have values for properties which are not editable (required and hidden or read-only). This fixes a bug where the user was not able to save a form, because he could not fill out a required value.
Converter for deserializing Dictionaries from HAL-Forms (KeyValueDictionaryConverter).